requirements.txt
# Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0

# Authentication (JWT)
djangorestframework-simplejwt>=5.3,<6.0

# Filtering & Search
django-filter>=24.0,<25.0

# CORS
django-cors-headers>=4.4,<5.0

# Environment Variables
python-decouple>=3.8,<4.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
faker>=24.0,<25.0

# Code Quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0
